rename: weather type 파일 이동 및 경로 수정
