+++ src/entities/weather/index.ts
@@ -4,4 +4,10 @@ export { WEATHER_TYPES } from './consts/weatherType';
 export { WEATHER_LABELS } from './consts/label';
 export { WEATHER_COLORS } from './consts/color';
 
-export type { WeatherTypeName, WeatherDto } from './model/types';
+export type {
+  WeatherType,
+  WeatherTypeName,
+  WeatherDto,
+  RainType,
+  SkyStatus,
+} from './model/types';
