+++ src/entities/weather/model/types.ts
@@ -1,5 +1,4 @@
-import { RainType, SkyStatus } from '@/shared/types';
-
+export type WeatherType = '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8';
 export type WeatherTypeName =
   | 'sweltering'
   | 'hot'
@@ -9,6 +8,18 @@ export type WeatherTypeName =
   | 'chilly'
   | 'cold'
   | 'frigid';
+
+export type RainType =
+  | 'NONE'
+  | 'RAIN'
+  | 'RAIN_AND_SNOW'
+  | 'SNOW'
+  | 'SHOWER'
+  | 'RAIN_DROP'
+  | 'RAIN_AND_SNOW_FLURRIES'
+  | 'SNOW_FLURRIES';
+
+export type SkyStatus = 'CLEAR' | 'PARTLY_CLOUDY' | 'CLOUDY';
 
 export type WeatherDto = {
   season: string;
