+++ src/pages/UserLookbookCreate/WeatherHeadline/WeatherHeadline.tsx
@@ -2,9 +2,9 @@ import { memo } from 'react';
 import { useLocation } from 'react-router-dom';
 
 import { TempCondition } from '@/entities/member/model/types';
+import { WeatherType } from '@/entities/weather';
 
 import { LOOKBOOK_WEATHER_TYPE } from '@/shared/consts';
-import { WeatherType } from '@/shared/types';
 
 import { LocationState } from '../ui/Page/UserLookbookCreatePage';
