+++ src/pages/UserLookbookCreate/ui/Page/UserLookbookCreatePage.tsx
@@ -3,9 +3,9 @@ import { Navigate, useLocation, useSearchParams } from 'react-router-dom';
 
 import { MemberLookbookDto } from '@/entities/clothes/model/types';
 import { TempCondition } from '@/entities/member/model/types';
+import { WeatherType } from '@/entities/weather';
 
 import { DEFAULT_CLOTHES_BY_WEATHER } from '@/shared/consts/lookbook';
-import { WeatherType } from '@/shared/types';
 import { HeadHelmet } from '@/shared/ui';
 
 import EditSection from '../../Edit/EditSection';
