--- src/shared/types/weather.ts
@@ -1,13 +1,0 @@
-export type RainType =
-  | 'NONE'
-  | 'RAIN'
-  | 'RAIN_AND_SNOW'
-  | 'SNOW'
-  | 'SHOWER'
-  | 'RAIN_DROP'
-  | 'RAIN_AND_SNOW_FLURRIES'
-  | 'SNOW_FLURRIES';
-
-export type SkyStatus = 'CLEAR' | 'PARTLY_CLOUDY' | 'CLOUDY';
-
-export type WeatherType = '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8';
